refactor(prompts): rename generic to custom to align naming convection
