+++ admin-client/src/components/prompts/custom/CheckboxListPrompt.vue
