+++ admin-client/src/components/prompts/custom/InfoPrompt.vue
