+++ common/src/prompts/props/index.ts
@@ -1,4 +1,4 @@
 export * from './base';
-export * from './generic';
+export * from './custom';
 export * from './portion-size';
 export * from './standard';
