+++ survey-client/src/components/prompts/custom/CheckboxListPrompt.vue
